newsapi/requirements.txt
APScheduler==3.7.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
coverage==5.5
Django==3.2.6
django-apscheduler==0.6.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-node==4.0.2
django-nose==1.4.7
djangorestframework==3.12.4
httpie==2.4.0
idna==3.2
nose==1.3.7
psycopg2==2.9.1
Pygments==2.9.0
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
schedule==1.1.0
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.6
